requirements.txt
hashlib==1.0.1
torch==1.9.1
pyro-ppl==1.7.0
numpy==1.21.2
pandas==1.3.3
yfinance==0.1.63
yahoofinancials==1.6
pandas-datareader==0.10.0
tqdm==4.62.3
torch==1.9.1
scikit-learn==0.24.2
